+++ web/app/themes/brocooly/index.php
@@ -19,6 +19,5 @@
  *
  * Brocooly handle requests with Route class - no other files except for functions.php required.
  * If you wish you may use any of them according to hierarchy (like singular.php) - it has higher priority.
- * But you should do not include TemplateRedirect hook than as it is redirect all queries into index.php since version 0.8.10.
  */
 app()->web();